Hackathon/backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[argon2]
groq
python-dotenv
sqlalchemy
python-multipart
pydantic[email]
email-validator
# AI and Search Dependencies
langchain-groq
langchain-community
duckduckgo-search
langchain
numpy
pillow
opencv-python-headless
deepface
fer
tf-keras
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
torchvision>=0.15.0 --index-url https://download.pytorch.org/whl/cpu
joblib
pandas
scikit-learn
python-doctr
tf2jax
tf-keras
